databricks/databricks-streamlit-sql-warehouse/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export -o requirements.txt --no-hashes --no-dev
altair==5.5.0
    # via streamlit
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via streamlit
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   loguru
databricks-sdk==0.52.0
    # via databricks-streamlit-sql-warehouse
databricks-sql-connector==4.0.3
    # via databricks-streamlit-sql-warehouse
et-xmlfile==2.0.0
    # via openpyxl
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-auth==2.40.0
    # via databricks-sdk
idna==3.10
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
loguru==0.7.3
    # via databricks-streamlit-sql-warehouse
lz4==4.4.4
    # via databricks-sql-connector
markupsafe==3.0.2
    # via jinja2
narwhals==1.38.0
    # via altair
numpy==2.2.5
    # via
    #   pandas
    #   pydeck
    #   streamlit
oauthlib==3.2.2
    # via databricks-sql-connector
openpyxl==3.1.5
    # via databricks-sql-connector
packaging==24.2
    # via
    #   altair
    #   posit-sdk
    #   streamlit
pandas==2.2.3
    # via
    #   databricks-sql-connector
    #   streamlit
pillow==11.2.1
    # via streamlit
polars==1.29.0
    # via databricks-streamlit-sql-warehouse
posit-sdk==0.9.0
    # via databricks-streamlit-sql-warehouse
protobuf==6.30.2
    # via streamlit
pyarrow==20.0.0
    # via
    #   databricks-streamlit-sql-warehouse
    #   streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via
    #   databricks-sql-connector
    #   pandas
pytz==2025.2
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   databricks-sdk
    #   databricks-sql-connector
    #   posit-sdk
    #   streamlit
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via
    #   python-dateutil
    #   thrift
smmap==5.0.2
    # via gitdb
streamlit==1.45.0
    # via databricks-streamlit-sql-warehouse
tenacity==9.1.2
    # via streamlit
thrift==0.20.0
    # via databricks-sql-connector
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typing-extensions==4.13.2
    # via
    #   altair
    #   posit-sdk
    #   referencing
    #   streamlit
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via
    #   databricks-sql-connector
    #   requests
watchdog==6.0.0 ; sys_platform != 'darwin'
    # via streamlit
win32-setctime==1.2.0 ; sys_platform == 'win32'
    # via loguru
